requirements-dev.txt
# ONE3TAP - Development Requirements
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
factory-boy>=3.3.0
pytest-django>=4.7.0
coverage>=7.3.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
pre-commit>=3.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
